cafTestApplication: Move PDM object classes into PdmObjects subfolder
