+++ Fwk/AppFwk/cafTests/cafTestApplication/PdmObjects/ColorTriplet.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#include "cafPdmField.h"
+#include "cafPdmObject.h"
+
+#include <QString>
+
+class ColorTriplet : public caf::PdmObject
+{
+    CAF_PDM_HEADER_INIT;
+
+public:
+    ColorTriplet();
+
+private:
+    caf::PdmField<QString> m_colorField;
+    caf::PdmField<int>     m_categoryValue;
+    caf::PdmField<QString> m_categoryText;
+};
