+++ Fwk/AppFwk/cafTests/cafTestApplication/PdmObjects/DemoPdmObjectGroup.cpp
@@ -1,0 +1,10 @@
+#include "DemoPdmObjectGroup.h"
+
+CAF_PDM_SOURCE_INIT( DemoPdmObjectGroup, "DemoPdmObjectGroup" );
+
+DemoPdmObjectGroup::DemoPdmObjectGroup()
+{
+    CAF_PDM_InitFieldNoDefault( &objects, "PdmObjects", "MyRootObject", "", "", "" );
+
+    objects.uiCapability()->setUiHidden( true );
+}
