+++ Fwk/AppFwk/cafTests/cafTestApplication/PdmObjects/DemoPdmObjectGroup.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "cafPdmChildArrayField.h"
+#include "cafPdmDocument.h"
+#include "cafPdmObject.h"
+
+class DemoPdmObjectGroup : public caf::PdmDocument
+{
+    CAF_PDM_HEADER_INIT;
+
+public:
+    DemoPdmObjectGroup();
+
+public:
+    caf::PdmChildArrayField<caf::PdmObjectHandle*> objects;
+};
